[FL-2520] FW build with -Wextra  (#1185)
* Fixing compiler warnings with -Wextra
* More warnings suppression, WIP
* Even more warning fixes
* Added new lines at end of text files.
* Padding fix
* Additional fixes to warnings on different build configurations; added -Wextra to default build pipeline
* Fixes for Secplus v1
* -additional warnings
* +-Wredundant-decls fixes
* FuriHal: print stack overflow task name in console
* FuriHal: add missing include

+++ lib/infrared/encoder_decoder/common/infrared_common_encoder.c
@@ -146,7 +146,7 @@ void* infrared_common_encoder_alloc(const InfraredCommonProtocolSpec* protocol) 
     }
 
     /* protocol->databit_len[0] has to contain biggest value of bits that can be decoded */
-    for(int i = 1; i < COUNT_OF(protocol->databit_len); ++i) {
+    for(size_t i = 1; i < COUNT_OF(protocol->databit_len); ++i) {
         furi_assert(protocol->databit_len[i] <= protocol->databit_len[0]);
     }
 
@@ -174,7 +174,7 @@ void infrared_common_encoder_reset(InfraredCommonEncoder* encoder) {
 
     uint8_t max_databit_len = 0;
 
-    for(int i = 0; i < COUNT_OF(encoder->protocol->databit_len); ++i) {
+    for(size_t i = 0; i < COUNT_OF(encoder->protocol->databit_len); ++i) {
         max_databit_len = MAX(max_databit_len, encoder->protocol->databit_len[i]);
     }
